+++ types/components/initClipboard.d.ts
@@ -1,5 +1,5 @@
 /**
  * Initialize the clipboard button.
  */
-declare const _default: (context: JQuery | HTMLElement, delayMSec?: number) => void;
+declare const _default: (context: JQuery | HTMLElement, delay?: number) => void;
 export default _default;